requirements.txt
Flask==2.2.2
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
ydata-profiling==4.1.2
xhtml2pdf==0.2.11
google-generativeai==0.3.2